+++ iod/src/cw.cpp
@@ -610,8 +610,8 @@ int main(int argc, char const *argv[]) {
                   << "\nconst char *name_from_id(int id);\n";
             // export messages
             {
-                std::map<std::string, int> &messages = ExportState::all_messages();
-                std::map<std::string, int>::const_iterator iter = messages.begin();
+                std::map<std::string, size_t> &messages = ExportState::all_messages();
+                std::map<std::string, size_t>::const_iterator iter = messages.begin();
                 while (iter != messages.end()) {
                     const std::pair<std::string, int> &item = *iter++;
                     if (item.second >=
@@ -629,8 +629,8 @@ int main(int argc, char const *argv[]) {
             std::ofstream msg_c(msg_support);
             msg_c << "\nconst char *name_from_id(int id) {\n";
             {
-                std::map<std::string, int> &messages = ExportState::all_messages();
-                std::map<std::string, int>::const_iterator iter = messages.begin();
+                std::map<std::string, size_t> &messages = ExportState::all_messages();
+                std::map<std::string, size_t>::const_iterator iter = messages.begin();
                 while (iter != messages.end()) {
                     const std::pair<std::string, int> &item = *iter++;
                     msg_c << "\tif (id==" << item.second << ") return \"" << item.first << "\";\n";
@@ -649,8 +649,8 @@ int main(int argc, char const *argv[]) {
                   << "#define __cw_symbol_ids_h__\n\n";
             // export messages
             {
-                std::map<std::string, int> &symbols = ExportState::all_symbols();
-                std::map<std::string, int>::const_iterator iter = symbols.begin();
+                std::map<std::string, size_t> &symbols = ExportState::all_symbols();
+                std::map<std::string, size_t>::const_iterator iter = symbols.begin();
                 while (iter != symbols.end()) {
                     const std::pair<std::string, int> &item = *iter++;
                     msg_h << "#define " << item.first << " " << item.second << "\n";
